Fixes NEMO#272: SelectionDialog doesn't work with C++ model out of the box.
model.count relied upon the model having a count property, likewise
model.length required a length property. Out of the box, C++ models don't have
these properties, and the behaviour for not having them is very unintuitive: it
just stops working.

ListView.count is a C++-implemented property that wraps all the different types
of models' differences so that we don't have to deal with this. We're already
using its notify signal, so why not just use ListView.count to calculate our
data, too? :-)