sectionscroller: better cope with custom model types.
By using ListView::count, we can avoid all of the stupid fiddling
to determine what property models are using to expose their length
this week.

This also avoids the need to expose count/length properties in C++
models, which is an major pain with proxy models.

As an added bonus, we no longer need to explicitly connect to
rowsInserted/rowsRemoved, because we'll see those implicitly with
the model length changing.