fix obnoxious bug in sheet placement, seen in e.g. accounts-ui
for whatever reason, this was frequently failing (especially with lazy-
loading of sheets) to detect the parent, causing them to incorrectly
apply a status bar-sized margin to the top of the sheet.

take advantage of the fact that parent is always set or detected,
and use the same page detection we use to find it in the first place.