Hide the status bar if there is no valid pixmap to be drawn
Adds a hasPixmap property to the MDeclarativeStatusBar which can be used to hide the status bar when there is no valid pixmap.
There is no need to request the pixmap on every paint() - there's a D-Bus service watcher for the pixmap service in place so querying the pixmap when the service provider changes is enough.