Clean up rotation angle property setting.
This code was not only messy, it's all kinds of broken. It reset the window id
it used every time a window was mapped, which is useful for supporting multiple
windows (which admittedly needs investigation), but broke things terribly in
that after a map, it would always update the _wrong_ window.

This caused a bug in e.g.
https://github.com/thp/lipstick-colorful-home/commit/be7a676e8c0513a294f4fc9d4318942ca7be98a4
where something would be mapped, the window id would be reset, and thus the
orientation angle for the desktop would never correctly be set again, meaning
sysuid would show on the wrong orientation.

As another side effect, this should mean that all processes no longer wake up
and set their rotation angle (uselessly) on a window being mapped.