[localthemedaemonclient] PNGs are now preferred over SVGs
Theme inheritance does not work well with file types.
Earlier, if base has x.svg, it will be preferred over darko's x.png
This now has been changed, to prefer PNGs, but there's still that bug
not inheriting properly.
It is deemed safe to go this way for now, as rendering SVGs is slow